+++ packages/crypto-transaction/source/index.ts
@@ -4,7 +4,7 @@ export * from "./factory.js";
 export * from "./serializer.js";
 export * from "./service-provider.js";
 export * from "./signer.js";
-export * from "./types/index.js";
+export * from "./transaction.js";
 export * from "./utilities.js";
 export * from "./validation/index.js";
 export * from "./verifier.js";
