--- packages/crypto-transaction/source/types/index.ts
@@ -1,1 +1,0 @@
-export * from "./transaction.js";
